requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\Clement\Documents\GitHub\necrobot\rings\meta.py: 11
# C:\Users\Clement\Documents\GitHub\necrobot\rings\profile.py: 12,13,14
Pillow >= 8.2.0

# C:\Users\Clement\Documents\GitHub\necrobot\rings\meta.py: 8
# C:\Users\Clement\Documents\GitHub\necrobot\rings\misc.py: 11
# C:\Users\Clement\Documents\GitHub\necrobot\rings\utilities.py: 8
# C:\Users\Clement\Documents\GitHub\necrobot\rings\words.py: 9
aiohttp >= 3.7.4.post0

# C:\Users\Clement\Documents\GitHub\necrobot\rings\db.py: 6
asyncpg >= 0.22.0

# C:\Users\Clement\Documents\GitHub\necrobot\rings\bridge.py: 12
# C:\Users\Clement\Documents\GitHub\necrobot\rings\modding.py: 5
# C:\Users\Clement\Documents\GitHub\necrobot\rings\rss.py: 10
# C:\Users\Clement\Documents\GitHub\necrobot\rings\social.py: 9
beautifulsoup4 >= 4.6.3

# C:\Users\Clement\Documents\GitHub\necrobot\rings\decisions.py: 7
dice >= 3.1.2

# C:\Users\Clement\Documents\GitHub\necrobot\bot.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\maintenance_bot.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\admin.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\bridge.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\db.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\decisions.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\economy.py: 2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\events.py: 2,3
# C:\Users\Clement\Documents\GitHub\necrobot\rings\meta.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\misc.py: 1,2,3
# C:\Users\Clement\Documents\GitHub\necrobot\rings\modding.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\moderation.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\profile.py: 1,2,3
# C:\Users\Clement\Documents\GitHub\necrobot\rings\rss.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\server.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\social.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\support.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\tags.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\utilities.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\utils\converters.py: 1,2,3
# C:\Users\Clement\Documents\GitHub\necrobot\rings\utils\cooldowns.py: 1
# C:\Users\Clement\Documents\GitHub\necrobot\rings\utils\help.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\utils\utils.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\wiki.py: 1,2
# C:\Users\Clement\Documents\GitHub\necrobot\rings\words.py: 1,2,3
discord.py >= 2.3.0

# C:\Users\Clement\Documents\GitHub\necrobot\rings\server.py: 9
emoji >= 1.2.0

# C:\Users\Clement\Documents\GitHub\necrobot\rings\rss.py: 6
feedparser >= 5.2.1

# C:\Users\Clement\Documents\GitHub\necrobot\rings\modding.py: 6
# C:\Users\Clement\Documents\GitHub\necrobot\rings\wiki.py: 8
fuzzywuzzy >= 0.18.0

# C:\Users\Clement\Documents\GitHub\necrobot\rings\modding.py: 7
git+https://github.com/ClementJ18/moddb.git

# C:\Users\Clement\Documents\GitHub\necrobot\rings\admin.py: 11
pbs >= 0.110

# C:\Users\Clement\Documents\GitHub\necrobot\rings\admin.py: 15
psutil >= 5.8.0

# C:\Users\Clement\Documents\GitHub\necrobot\rings\db.py: 7
psycopg2_binary >= 2.8.6

# C:\Users\Clement\Documents\GitHub\necrobot\rings\bridge.py: 13
robobrowser >= 0.5.3

# C:\Users\Clement\Documents\GitHub\necrobot\rings\admin.py: 18
# C:\Users\Clement\Documents\GitHub\necrobot\rings\utilities.py: 10
simpleeval >= 0.9.10

yarl >= 1.4.2
pytz >= 2021.03
Werkzeug == 0.16.1
python-Levenshtein >= 0.12.2
black == 22.3.0
flake8 == 3.9.0
flakehell == 0.9.0
pathfinding == 1.0.1
pandas==2.0.2
matplotlib==3.7.1
scipy==1.10.1
